reorg: move bootState16 to bootstate16.go for clarity
